fix: Improved logic for Ascending and Descending from the Aether text, this should fix it appearing in places it shouldn't.
