+++ src/main/java/com/gildedgames/aether/common/event/listeners/DimensionListener.java
@@ -49,13 +49,6 @@ public class DimensionListener {
     }
 
     @SubscribeEvent
-    public static void onEntityTravelToDimension(EntityTravelToDimensionEvent event) {
-        Entity entity = event.getEntity();
-        ResourceKey<Level> dimension = event.getDimension();
-        DimensionHooks.dimensionTravel(entity, dimension);
-    }
-
-    @SubscribeEvent
     public static void onWorldTick(TickEvent.WorldTickEvent event) {
         Level level = event.world;
         if (event.side == LogicalSide.SERVER && event.phase == TickEvent.Phase.END) {
